- Encapsulate file logging into LogUtil - Add a button to view detailed error log to the "retry download" popup - Basic error logging during download

+++ app/src/main/java/me/devsaki/hentoid/database/HentoidDB.java
@@ -141,7 +141,7 @@ public class HentoidDB extends SQLiteOpenHelper {
     }
 
     @Nullable
-    public Content selectContentById(int id) {
+    public Content selectContentById(long id) {
         Content result;
         Timber.d("selectContentById");
         SQLiteDatabase db = openDatabase();
